requirements.txt
requests==2.29.0
beautifulsoup4
numpy
pandas
scikit-learn
joblib
Pillow
flask
gunicorn
tensorflow>=2.21.0,<2.22